+++ src/main.cpp
@@ -9,7 +9,7 @@ public:
     std::vector<std::unordered_map<int, int>> weights;
     std::vector<std::vector<int>> g;
 
-    int get_cost(int a, int b) override {
+    double get_cost(int a, int b) override {
         return weights[a][b];
     }
